[rename] 프로젝트 서비스 링크 컴포넌트 파일명 변경
